+++ src/main/java/com/soda/article/dto/article/ArticleModifyRequest.java
@@ -1,4 +1,4 @@
-package com.soda.article.domain.article;
+package com.soda.article.dto.article;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.soda.article.enums.PriorityType;
