+++ src/main/java/com/soda/article/dto/article/ArticleViewResponse.java
@@ -1,4 +1,4 @@
-package com.soda.article.domain.article;
+package com.soda.article.dto.article;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.soda.article.entity.Article;
