+++ src/main/java/com/soda/article/dto/comment/CommentCreateRequest.java
@@ -1,4 +1,4 @@
-package com.soda.article.domain.comment;
+package com.soda.article.dto.comment;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
