+++ src/main/java/com/soda/article/service/ArticleService.java
@@ -1,6 +1,6 @@
 package com.soda.article.service;
 
-import com.soda.article.domain.article.*;
+import com.soda.article.dto.article.*;
 import com.soda.article.entity.Article;
 import com.soda.article.entity.ArticleLink;
 import com.soda.article.error.ArticleErrorCode;
