+++ src/main/java/com/soda/global/log/dataLog/aspect/EntityLogAspect.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.soda.article.domain.article.ArticleViewResponse;
-import com.soda.article.domain.comment.CommentDTO;
+import com.soda.article.dto.article.ArticleViewResponse;
+import com.soda.article.dto.comment.CommentDTO;
 import com.soda.article.entity.Article;
 import com.soda.article.entity.Comment;
 import com.soda.global.log.dataLog.annotation.LoggableEntityAction;
